Make setup of projection in Camera Virtual, and the ProjectionMatrix property protected so Camera can be derived and modified, also implemented a OrthographicCamera - allthough not exactly working as I thought it would - need to look into this more.

+++ Source/Balder.Core/Math/Frustum.cs
@@ -84,7 +84,7 @@ namespace Balder.Core.Math
 		{
 			Vector nc, fc, X, Y, Z;
 
-			SetCameraInternals(camera.FieldOfView, viewport.AspectRatio, camera.Near, camera.Far);
+			SetCameraInternals((float)camera.FieldOfView, viewport.AspectRatio, camera.Near, camera.Far);
 
 			Z = camera.Target-camera.Position;
 			Z.Normalize();
